+++ compile/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./vite-dynamic-comments";
+export * from "./css-imported";