+++ packages/vitdoc/CHANGELOG.md
@@ -1,5 +1,15 @@
 # vitdoc
 
+## 1.4.14
+
+### Patch Changes
+
+- fix: vite esm module css load
+- Updated dependencies
+  - @vitdoc/runtime@1.4.8
+  - @vitdoc/ui@1.3.7
+  - @vitdoc/theme-default@2.0.7
+
 ## 1.4.13
 
 ### Patch Changes